+++ src/components/Menu/Menu/__stories__/menu.stories.js
@@ -35,7 +35,7 @@ export const menuTemplate = args => (
 
 export const menuSizesTemplate = args => [
   <DialogContentContainer key="small">
-    <Menu size={Menu.sizes.SMALL}>
+    <Menu {...args} size={Menu.sizes.SMALL}>
       <MenuTitle caption="Small menu" />
       <MenuDivider />
       <MenuItem title="Menu item 1" />
@@ -44,7 +44,7 @@ export const menuSizesTemplate = args => [
     </Menu>
   </DialogContentContainer>,
   <DialogContentContainer key="md">
-    <Menu size={Menu.sizes.MEDIUM}>
+    <Menu {...args} size={Menu.sizes.MEDIUM}>
       <MenuTitle caption="Medium menu" />
       <MenuDivider />
       <MenuItem title="Menu item 1" />
@@ -53,7 +53,7 @@ export const menuSizesTemplate = args => [
     </Menu>
   </DialogContentContainer>,
   <DialogContentContainer key="lg">
-    <Menu size={Menu.sizes.LARGE}>
+    <Menu {...args} size={Menu.sizes.LARGE}>
       <MenuTitle caption="Large menu" />
       <MenuDivider />
       <MenuItem title="Menu item 1" />
@@ -75,7 +75,7 @@ export const menuWithIconsTemplate = args => (
 
 export const menuWithSubMenuTemplate = args => (
   <DialogContentContainer>
-    <Menu>
+    <Menu {...args}>
       <MenuItem title="Menu item" icon={Activity} />
       <MenuItem title='Hover me to see the sub menu"' icon={Activity}>
         <Menu>
@@ -91,7 +91,7 @@ export const menuWithSubMenuTemplate = args => (
 
 export const menuWith2DepthSubMenuTemplate = args => (
   <DialogContentContainer>
-    <Menu>
+    <Menu {...args}>
       <MenuItem title="Menu item" icon={Favorite} />
       <MenuItem title="Hover me to see the sub menu" icon={Activity}>
         <Menu>
@@ -114,7 +114,7 @@ export const menuWith2DepthSubMenuTemplate = args => (
 export const menuWithGridItems = args => (
   <div className={classes["menu-long-story-wrapper"]}>
     <DialogContentContainer>
-      <Menu>
+      <Menu {...args}>
         <MenuItem title="Menu item" icon={Favorite} />
         <MenuTitle caption="Top level grid item" />
         <MenuItem title="Hover me to see the sub menu" icon={Activity}>